move define_syscall macro to its own crate
