+++ define-syscall/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "solana-define-syscall"
+description = "Solana define_syscall macro. For internal use only."
+documentation = "https://docs.rs/solana-define-syscall"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
